data_pipeline/scrapers/conference_scraper.py
"""
Conference Data Scraper
Scrapes data from SOT, AACR, ACT conferences

This module extracts presenter information from conference abstracts and programs.
"""

import requests
from typing import List, Dict, Any, Optional
from datetime import datetime
import logging
from bs4 import BeautifulSoup

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class ConferenceScraper:
    """
    Scraper for conference data (SOT, AACR, ACT).
    """
    
    def __init__(self):
        """Initialize conference scraper."""
        self.conference_sources = {
            "SOT": "https://www.toxicology.org",
            "AACR": "https://www.aacr.org",
            "ACT": "https://www.actox.org"
        }
        self.rate_limit_delay = 1.0
    
    def search_conference_presenters(
        self,
        conference: str,
        keywords: List[str],
        year: Optional[int] = None
    ) -> List[Dict[str, Any]]:
        """
        Search for conference presenters matching keywords.
        
        Args:
            conference: One of "SOT", "AACR", "ACT"
            keywords: List of keywords to search for
            year: Conference year (default: current year)
        
        Returns:
            List of presenter dictionaries
        """
        try:
            if year is None:
                year = datetime.now().year
            
            if conference.upper() == "SOT":
                return self._search_sot(keywords, year)
            elif conference.upper() == "AACR":
                return self._search_aacr(keywords, year)
            elif conference.upper() == "ACT":
                return self._search_act(keywords, year)
            else:
                logger.warning(f"Unknown conference: {conference}")
                return []
        
        except Exception as e:
            logger.error(f"Error searching conference data: {str(e)}")
            logger.error("Conference scraping error - returning empty results.")
            return []
    
    def _search_sot(
        self,
        keywords: List[str],
        year: int
    ) -> List[Dict[str, Any]]:
        """Search SOT (Society of Toxicology) conference data."""
        # SOT abstracts are typically available at:
        # https://www.toxicology.org/abstracts/search.asp
        # Implementation would require web scraping of their abstract database
        logger.warning(f"SOT {year} scraping not yet implemented. Conference web scraping required.")
        return []
    
    def _search_aacr(
        self,
        keywords: List[str],
        year: int
    ) -> List[Dict[str, Any]]:
        """Search AACR (American Association for Cancer Research) conference data."""
        # AACR abstracts available at their website
        # Implementation would require web scraping
        logger.warning(f"AACR {year} scraping not yet implemented. Conference web scraping required.")
        return []
    
    def _search_act(
        self,
        keywords: List[str],
        year: int
    ) -> List[Dict[str, Any]]:
        """Search ACT (American College of Toxicology) conference data."""
        # ACT conference data available at their website
        # Implementation would require web scraping
        logger.warning(f"ACT {year} scraping not yet implemented. Conference web scraping required.")
        return []


def search_conferences(
    conferences: List[str],
    keywords: List[str],
    year: Optional[int] = None
) -> List[Dict[str, Any]]:
    """
    Search multiple conferences for presenters.
    
    Args:
        conferences: List of conference names
        keywords: Search keywords
        year: Conference year
    
    Returns:
        Combined list of presenters
    """
    scraper = ConferenceScraper()
    all_presenters = []
    
    for conference in conferences:
        presenters = scraper.search_conference_presenters(conference, keywords, year)
        all_presenters.extend(presenters)
    
    return all_presenters


if __name__ == "__main__":
    # Test the scraper
    scraper = ConferenceScraper()
    results = scraper.search_conference_presenters(
        conference="SOT",
        keywords=["toxicology", "3D models", "DILI"],
        year=2024
    )
    print(f"Found {len(results)} presenters")
    for presenter in results:
        print(f"- {presenter['name']}: {presenter['presentation_title']}")

