requirements.txt
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-search-results>=2.4.2
python-dotenv>=1.0.0

